requirements.txt
adafruit-circuitpython-sht4x
paho-mqtt
pyserial
